fix bug of --recordsmem option
git-svn-id: http://192.168.20.11/sequoiadb/trunk@19954 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/import/impOptions.cpp
@@ -178,7 +178,7 @@ namespace import
 
       _bufferSize = 64;
       _dryRun = FALSE;
-      _recordsMem = (1024 * 1024 * 1024 * 2); // 2GB
+      _recordsMem = (INT64)1024 * 1024 * 1024 * 2; // 2GB
    }
 
    Options::~Options()
